refactor(rod): refactor ETL model layer
