+++ rod/etl/models/customer.py
@@ -1,0 +1,47 @@
+import uuid
+
+from django.conf import settings
+from django.contrib import admin
+from django.db import models
+
+from rod.common.models import BaseModel
+from rod.etl.validators import UserValidator
+
+PHONE_LENGTH = 10
+
+
+# Create your models here.
+class Customer(BaseModel):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    phone = models.CharField(max_length=PHONE_LENGTH, default="", blank=True)
+    address = models.TextField(default="", blank=True)
+    birth_date = models.DateField(null=True, blank=True)
+    image = models.ImageField(
+        upload_to="customer/images",
+        blank=True,
+        null=True,
+    )
+    user = models.OneToOneField(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        db_index=True,
+    )
+
+    def __str__(self):
+        return f"{self.user.first_name} {self.user.last_name}"
+
+    @admin.display(ordering="user__first_name")
+    def first_name(self):
+        return self.user.first_name
+
+    @admin.display(ordering="user__last_name")
+    def last_name(self):
+        return self.user.last_name
+
+    def clean(self):
+        UserValidator().validate_birth_date(self.birth_date)
+        UserValidator().validate_phone(self.phone)
+        UserValidator().validate_file_size(self.image)
+
+    class Meta:
+        ordering = ["user__first_name", "user__last_name"]
