+++ rod/etl/selectors/employee.py
@@ -1,0 +1,13 @@
+import uuid
+
+from rod.common.utils import get_object
+from rod.etl.models.employee import Employee
+
+
+class EmployeeSelector:
+    def __init__(self) -> None:
+        pass
+
+    def employee_get(self, *, user_id: uuid.UUID) -> Employee | None:
+        queryset = Employee.objects.select_related("user")
+        return get_object(queryset, user_id=user_id)
