+++ rod/etl/services/customer.py
@@ -1,6 +1,5 @@
 from rod.common.services import model_update
-from rod.etl.models import Customer
-from rod.etl.models import Employee
+from rod.etl.models.customer import Customer
 from rod.users.models import BaseUser
 
 
@@ -27,28 +26,3 @@ class CustomerService:
             data=data,
         )
         return customer
-
-
-class EmployeeService:
-    def __init__(self) -> None:
-        pass
-
-    def employee_create(self, *, user: BaseUser) -> Employee:
-        employee = Employee(user=user)
-        employee.full_clean()
-        employee.save()
-
-        return employee
-
-    def employee_update(self, *, employee: Employee, data) -> Employee:
-        fields: list[str] = [
-            "phone",
-            "address",
-            "birth_date",
-        ]
-        employee, has_updated = model_update(
-            instance=employee,
-            fields=fields,
-            data=data,
-        )
-        return employee
